+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2 (2017/10/14)
+
+* Username and avatar are embedded
+* Review post content
+
 ## 0.1 (2017/11/29)
 
 * Initial version